+++ openscm/errors.py
@@ -1,0 +1,47 @@
+class ParameterError(Exception):
+    """
+    Exception relating to a parameter. Used as super class.
+    """
+
+
+class ParameterLengthError(ParameterError):
+    """
+    Exception raised when sequences in timeseries do not match run
+    size.
+    """
+
+
+class ParameterReadonlyError(ParameterError):
+    """
+    Exception raised when a requested parameter is read-only.
+
+    This can happen, for instance, if a parameter's parent parameter
+    in the parameter hierarchy has already been requested as writable.
+    """
+
+
+class ParameterTypeError(ParameterError):
+    """
+    Exception raised when a parameter is of a different type than
+    requested (scalar or timeseries).
+    """
+
+
+class ParameterAggregatedError(ParameterError):
+    """
+    Exception raised when a parameter has been read from (raised, e.g., when attempting
+    to create a child parameter).
+    """
+
+
+class ParameterWrittenError(ParameterError):
+    """
+    Exception raised when a parameter has been written to (raised, e.g., when attempting
+    to create a child parameter).
+    """
+
+
+class RegionAggregatedError(Exception):
+    """
+    Exception raised when a region has been read from in a region-aggregated way.
+    """
