+++ openscm/regions.py
@@ -1,0 +1,141 @@
+from typing import Dict, Tuple
+from .errors import RegionAggregatedError
+from .parameters import _Parameter
+
+class _Region:
+    """
+    Represents a region in the region hierarchy.
+    """
+
+    _children: Dict[str, "_Region"]
+    """Subregions"""
+
+    _has_been_aggregated: bool
+    """
+    If True, a parameter of this region has already been read in an aggregated way,
+    i.e., aggregating over subregions
+     """
+
+    _name: str
+    """Name"""
+
+    _parameters: Dict[str, _Parameter]
+    """Parameters"""
+
+    _parent: "_Region"
+    """Parent region (or `None` if root region)"""
+
+    def __init__(self, name: str):
+        """
+        Initialize
+
+        Parameters
+        ----------
+        name
+            Name
+        """
+        self._name = name
+        self._children = {}
+        self._has_been_aggregated = False
+        self._parameters = {}
+        self._parent = None
+
+    def get_or_create_subregion(self, name: str) -> "_Region":
+        """
+        Get a (direct) subregion of this region. Create and add it if not found.
+
+        Parameters
+        ----------
+        name
+            Name
+
+        Raises
+        ------
+        RegionAggregatedError
+            If the subregion would need to be added and a parameter of this region has
+            already been read in an aggregated way. In this case a subregion cannot be
+            created.
+        """
+        res = self._children.get(name, None)
+        if res is None:
+            if self._has_been_aggregated:
+                raise RegionAggregatedError
+            res = _Region(name)
+            res._parent = self
+            self._children[name] = res
+        return res
+
+    def get_subregion(self, name: Tuple[str]) -> "_Region":
+        """
+        Get a subregion of this region or ``None`` if not found.
+
+        Parameters
+        ----------
+        name
+            Hierarchical name of the region below this region or ``()`` for this region.
+        """
+        if len(name) > 0:
+            res = self._children.get(name[0], None)
+            if res is not None:
+                res = res.get_subregion(name[1:])
+            return res
+        else:
+            return self
+
+    def get_or_create_parameter(self, name: str) -> _Parameter:
+        """
+        Get a root parameter for this region. Create and add it if not found.
+
+        Parameters
+        ----------
+        name
+            Name
+        """
+        res = self._parameters.get(name, None)
+        if res is None:
+            res = _Parameter(name)
+            self._parameters[name] = res
+        return res
+
+    def get_parameter(self, name: Tuple[str]) -> _Parameter:
+        """
+        Get a (root or sub-) parameter for this region or ``None`` if not found.
+
+        Parameters
+        ----------
+        name
+            :ref:`Hierarchical name <parameter-hierarchy>` of the parameter
+
+        Raises
+        ------
+        ValueError
+            Name not given
+        """
+        if len(name) > 0:
+            root_parameter = self._parameters.get(name, None)
+            if root_parameter is not None and len(name) > 1:
+                return root_parameter.get_subparameter(name[1:])
+            return root_parameter
+        else:
+            raise ValueError
+
+    def attempt_aggregate(self) -> None:
+        """
+        Tell region that one of its parameters will be read from in an aggregated way,
+        i.e., aggregating over subregions.
+        """
+        self._has_been_aggregated = True
+
+    @property
+    def name(self) -> str:
+        """
+        Name
+        """
+        return self._name
+
+    @property
+    def parent(self) -> "_Region":
+        """
+        Parent region (or ``None`` if root region)
+        """
+        return self._parent
