+++ tests/unit/test_core.py
@@ -1,12 +1,11 @@
-from openscm.core import (
+from openscm.errors import (
     ParameterAggregatedError,
     ParameterReadonlyError,
-    ParameterSet,
-    ParameterType,
     ParameterTypeError,
     ParameterWrittenError,
     RegionAggregatedError,
 )
+from openscm.core import ParameterSet, ParameterType
 import pytest
 
 
@@ -33,7 +32,7 @@ def test_region(parameterset):
 def test_parameter(parameterset):
     region_ber = parameterset._get_or_create_region(("DEU", "BER"))
 
-    with pytest.raises(ValueError, match="No region name given"):
+    with pytest.raises(ValueError, match="No parameter name given"):
         parameterset._get_or_create_parameter(
             (), region_ber, "GtCO2/a", ParameterType.TIMESERIES
         )
